# xe: conv: avoid reorder with NCHW when possible