Jira: [MFDNN-12923](https://jira.devtools.intel.com/browse/MFDNN-12923)

Performance data from BMG:

| Case                                                                                             | base perf (GOps/sec) | test perf (GOps/sec) | Ratio |
|--------------------------------------------------------------------------------------------------|----------------------|----------------------|-------|
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb128ic64ih64oc256oh64kh1ph0    | 3605.594             | 21190.681            | 5.88  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb64ic64ih128oc128oh128kh1ph0   | 3201.723             | 17190.661            | 5.37  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb256ic128ih72oc128oh72kh1ph0   | 5168.633             | 25795.306            | 4.99  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb128ic256ih64oc128oh64kh1ph0   | 7003.993             | 34561.331            | 4.93  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb256ic64ih144oc64oh144kh1ph0   | 2642.833             | 12802.686            | 4.84  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb128ic32ih128oc64oh128kh3ph1   | 14154.447            | 14187.520            | 1.00  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb256ic128ih72oc128oh72kh3ph1   | 37082.280            | 37114.164            | 1.00  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb64ic32ih128oc64oh128kh3ph1    | 14999.980            | 15005.511            | 1.00  |
| --dir=FWD_D --dt=bf16:bf16:bf16 --stag=abx --wtag=abx --dtag=abx mb64ic32ih256oc64oh128kh3sh2ph1 | 8451.155             | 8451.155             | 1.00  |


Improvements are expected for "good" sizes for 1x1 convolution with unit strides.