@echeresh 2025-04-04
make test perf-gpu
set primitive=conv deconv