@echeresh 2025-04-05
make test perf-gpu
set primitive=conv deconv