@echeresh 2025-04-08
make test perf-gpu
set primitive=conv deconv